#1539: Move multiple test case classes to their own package.
