+++ spotbugs-tests/src/test/java/edu/umd/cs/findbugs/detect/Issue1539Test.java
@@ -11,7 +11,7 @@ import static org.junit.Assert.assertThat;
 public class Issue1539Test extends AbstractIntegrationTest {
     @Test
     public void testInstance() {
-        performAnalysis("ghIssues/Issue1539Instance.class");
+        performAnalysis("ghIssues/issue1539/Issue1539Instance.class");
         BugInstanceMatcher bugTypeMatcher = new BugInstanceMatcherBuilder()
                 .bugType("DMI_RANDOM_USED_ONLY_ONCE")
                 .build();
@@ -20,7 +20,7 @@ public class Issue1539Test extends AbstractIntegrationTest {
 
     @Test
     public void testStatic() {
-        performAnalysis("ghIssues/Issue1539Static.class");
+        performAnalysis("ghIssues/issue1539/Issue1539Static.class");
         BugInstanceMatcher bugTypeMatcher = new BugInstanceMatcherBuilder()
                 .bugType("DMI_RANDOM_USED_ONLY_ONCE")
                 .build();
