+++ spotbugsTestCases/src/java/ghIssues/issue1539/Issue1539Static.java
@@ -1,4 +1,4 @@
-package ghIssues;
+package ghIssues.issue1539;
 
 import java.util.Random;
